logbook changes
Moved the bib file with references to the correct logbook folder. Deleted unnecessary template files due to not using the template. Added explanation for multiple days about the alpha + beta diversity calculations and the Sankey plot maker. Added some explanation about how our data was sequenced and some of the GitHub changes that were made for the layout of the directories.